Fixed build warnings
Sidenote: -Wconversion is an incredibly useful compiler flag for
gcc.  It warns you every time you implicitly convert types if
there's a chance of overflow/loss of precision.

This should fix the bug I saw on skybridge, where a size_t was
being implicitly cast to an int, but the int wasn't big enough.

+++ src/mpi/drivers/SC2016/gram_driver.cpp
@@ -63,7 +63,7 @@ int main(int argc, char* argv[])
   /////////////////////////
   // Check for bad input //
   /////////////////////////
-  if (proc_grid_dims->prod() != nprocs){
+  if ((int)proc_grid_dims->prod() != nprocs){
     if (rank==0) {
       std::cerr << "Processor grid dimensions do not multiply to nprocs" << std::endl;
       std::cout << "Processor grid dimensions: " << *proc_grid_dims << std::endl;
